Infrastructure to register and fetch undo action requests.
This infrasture provides a way to allow execution of undo actions.  One
might think that we can always execute undo actions on error or explicit
rollabck by user, however there are cases when that is not posssible.
For example, (a) if the system crash while doing operation, then after
startup, we need a way to perform undo actions; (b) If we get error while
performing undo actions.

Apart from this, when there are large rollback requests, then it is quite
inefficient to perform all the undo actions and then return control to
user.

To allow efficient execution of the undo actions, we create three queues
and a hash table for the rollback requests.  A Xid based priority queue
which will allow us to process the requests of older transactions and help
us to move oldesdXidHavingUnappliedUndo (this is a xid-horizon below which
all the transactions are visible) forward.  A size-based queue which will
help us to perform the rollbacks of larger aborts in a timely fashion so
that we don't get stuck while processing them during discard of the logs.
An error queue to hold the requests for transactions that failed to apply
its undo.  The rollback hash table is used to avoid duplicate undo requests
by backends and discard worker.

Amit Kapila and Kuntal Ghosh, design idea by Andres Freund.