Fix bug in zheap_undo_actions
We need to fix the full_xid as invalid transaxtion once the
rollback for the transaction has benn completed for this block.  Else,
the previous transaction might be rolled back wrongly when the rollback for
the current transaction has been performed.  Currently, we don't have the
block chain complete info here.  We should store it the undo payload for
the first undo record of this block by the current transaction.